+++ src/test/java/acceptance/toy1/bjs/UserOfToyComponent1.java
@@ -38,26 +38,26 @@ public class UserOfToyComponent1 extends BjsObject
         bjs.setProperty(bjsClass, "lastToy", bjs.putWrapped(value, ToyComponent1BjsExport.Wrapper.class));
     }
 
-    public static Integer add(Integer offset, Integer int1, Integer int2)
+    public static Long add(Long offset, Long int1, Long int2)
     {
-        return bjs.getInteger(bjs.call(bjsClass, "add", bjs.putPrimitive(offset), bjs.putPrimitive(int1),
+        return bjs.getLong(bjs.call(bjsClass, "add", bjs.putPrimitive(offset), bjs.putPrimitive(int1),
                 bjs.putPrimitive(int2)));
     }
 
-    public static ToyComponent1BjsExport getToy(Integer int1, Integer int2)
+    public static ToyComponent1BjsExport getToy(Long int1, Long int2)
     {
         return bjs.getWrapped(bjs.call(bjsClass, "getToy", bjs.putPrimitive(int1), bjs.putPrimitive(int2)));
     }
 
-    public static Integer getSum(ToyComponent1BjsExport toy1, ToyComponent1BjsExport toy2)
+    public static Long getSum(ToyComponent1BjsExport toy1, ToyComponent1BjsExport toy2)
     {
-        return bjs.getInteger(bjs.call(bjsClass, "getSum", bjs.putWrapped(toy1, ToyComponent1BjsExport.Wrapper.class),
+        return bjs.getLong(bjs.call(bjsClass, "getSum", bjs.putWrapped(toy1, ToyComponent1BjsExport.Wrapper.class),
                 bjs.putWrapped(toy2, ToyComponent1BjsExport.Wrapper.class)));
     }
 
-    public static Integer getSum2(ToyComponent2BjsExport toy1, ToyComponent1BjsExport toy2)
+    public static Long getSum2(ToyComponent2BjsExport toy1, ToyComponent1BjsExport toy2)
     {
-        return bjs.getInteger(bjs.call(bjsClass, "getSum2", bjs.putWrapped(toy1, ToyComponent2BjsExport.Wrapper.class),
+        return bjs.getLong(bjs.call(bjsClass, "getSum2", bjs.putWrapped(toy1, ToyComponent2BjsExport.Wrapper.class),
                 bjs.putWrapped(toy2, ToyComponent1BjsExport.Wrapper.class)));
     }
 
@@ -71,8 +71,8 @@ public class UserOfToyComponent1 extends BjsObject
         return bjs.getDouble(bjs.call(bjsClass, "piSum"));
     }
 
-    public static Integer additionalValue(ToyComponent2BjsExport toy)
+    public static Long additionalValue(ToyComponent2BjsExport toy)
     {
-        return bjs.getInteger(bjs.call(bjsClass, "additionalValue", bjs.putWrapped(toy, ToyComponent2BjsExport.Wrapper.class)));
+        return bjs.getLong(bjs.call(bjsClass, "additionalValue", bjs.putWrapped(toy, ToyComponent2BjsExport.Wrapper.class)));
     }
 }
